+++ packages/core/src/index.ts
@@ -4,7 +4,7 @@ import { build } from './build'
 export * from './config'
 export * from './build'
 export { CorePluginOptions, createPlugin, name } from './plugin'
-export * from './utils'
+export * from './utils2'
 export * from './types'
 export * from './managers'
 export * from './generators'
