+++ packages/core/src/utils2/getEncodedText.ts
