+++ packages/core/src/utils2/isPromise.ts
