+++ packages/core/src/utils2/isURL.ts
