+++ packages/core/src/utils2/read.ts
