+++ packages/core/src/utils2/renderTemplate.test.ts
