+++ packages/core/src/utils2/renderTemplate.ts
