+++ packages/core/src/utils2/write.ts
